chore: move xds client to tools
