+++ tools/xds-client/stream/client.go
